+++ pcs/lib/commands/node.py
@@ -4,24 +4,26 @@ from pcs.common import reports
 from pcs.common.reports.item import ReportItem
 from pcs.lib.cib.node import update_node_instance_attrs
 from pcs.lib.cib.tools import IdProvider
-from pcs.lib.env import LibraryEnvironment
+from pcs.lib.env import LibraryEnvironment, WaitType
 from pcs.lib.errors import LibraryError
 from pcs.lib.pacemaker.live import get_local_node_name
 from pcs.lib.pacemaker.state import ClusterState
 
 
 @contextmanager
-def cib_runner_nodes(lib_env: LibraryEnvironment, wait):
-    lib_env.ensure_wait_satisfiable(wait)
+def cib_runner_nodes(lib_env: LibraryEnvironment, wait: WaitType):
+    wait_timeout = lib_env.ensure_wait_satisfiable(wait)
     yield (
         lib_env.get_cib(),
         lib_env.cmd_runner(),
         ClusterState(lib_env.get_cluster_state()).node_section.nodes,
     )
-    lib_env.push_cib(wait=wait)
+    lib_env.push_cib(wait_timeout=wait_timeout)
 
 
-def standby_unstandby_local(lib_env, standby, wait=False):
+def standby_unstandby_local(
+    lib_env: LibraryEnvironment, standby, wait: WaitType = False
+):
     """
     Change local node standby mode
 
@@ -35,7 +37,9 @@ def standby_unstandby_local(lib_env, standby, wait=False):
     )
 
 
-def standby_unstandby_list(lib_env, standby, node_names, wait=False):
+def standby_unstandby_list(
+    lib_env: LibraryEnvironment, standby, node_names, wait: WaitType = False
+):
     """
     Change specified nodes standby mode
 
@@ -50,7 +54,9 @@ def standby_unstandby_list(lib_env, standby, node_names, wait=False):
     )
 
 
-def standby_unstandby_all(lib_env, standby, wait=False):
+def standby_unstandby_all(
+    lib_env: LibraryEnvironment, standby, wait: WaitType = False
+):
     """
     Change all nodes standby mode
 
@@ -64,7 +70,9 @@ def standby_unstandby_all(lib_env, standby, wait=False):
     )
 
 
-def maintenance_unmaintenance_local(lib_env, maintenance, wait=False):
+def maintenance_unmaintenance_local(
+    lib_env: LibraryEnvironment, maintenance, wait: WaitType = False
+):
     """
     Change local node maintenance mode
 
@@ -79,7 +87,7 @@ def maintenance_unmaintenance_local(lib_env, maintenance, wait=False):
 
 
 def maintenance_unmaintenance_list(
-    lib_env, maintenance, node_names, wait=False
+    lib_env: LibraryEnvironment, maintenance, node_names, wait: WaitType = False
 ):
     """
     Change specified nodes maintenance mode
@@ -98,7 +106,9 @@ def maintenance_unmaintenance_list(
     )
 
 
-def maintenance_unmaintenance_all(lib_env, maintenance, wait=False):
+def maintenance_unmaintenance_all(
+    lib_env: LibraryEnvironment, maintenance, wait: WaitType = False
+):
     """
     Change all nodes maintenance mode
 
@@ -120,7 +130,9 @@ def _create_maintenance_unmaintenance_dict(maintenance):
     return {"maintenance": "on" if maintenance else ""}
 
 
-def _set_instance_attrs_local_node(lib_env, attrs, wait):
+def _set_instance_attrs_local_node(
+    lib_env: LibraryEnvironment, attrs, wait: WaitType
+):
     if not lib_env.is_cib_live:
         # If we are not working with a live cluster we cannot get the local node
         # name.
@@ -140,7 +152,9 @@ def _set_instance_attrs_local_node(lib_env, attrs, wait):
         )
 
 
-def _set_instance_attrs_node_list(lib_env, attrs, node_names, wait):
+def _set_instance_attrs_node_list(
+    lib_env: LibraryEnvironment, attrs, node_names, wait: WaitType
+):
     with cib_runner_nodes(lib_env, wait) as (cib, dummy_runner, state_nodes):
         known_nodes = [node.attrs.name for node in state_nodes]
         report_list = []
@@ -158,7 +172,9 @@ def _set_instance_attrs_node_list(lib_env, attrs, node_names, wait):
             )
 
 
-def _set_instance_attrs_all_nodes(lib_env, attrs, wait):
+def _set_instance_attrs_all_nodes(
+    lib_env: LibraryEnvironment, attrs, wait: WaitType
+):
     with cib_runner_nodes(lib_env, wait) as (cib, dummy_runner, state_nodes):
         for node in [node.attrs.name for node in state_nodes]:
             update_node_instance_attrs(